+++ src/frontend/SettingsDialog.cpp
@@ -19,11 +19,13 @@ SettingsDialog::~SettingsDialog()
 
 void SettingsDialog::updateSettings() {
     ui->checkBoxPrevWin->setChecked(gSettings->getEnterKeyClosesPrevWin());
+    ui->checkBoxCandidateWinHorizontal->setChecked(gSettings->getCandidateWinHorizontal());
 }
 
 void SettingsDialog::on_buttonBox_accepted()
 {
     gSettings->setEnterKeyClosesPrevWin(ui->checkBoxPrevWin->isChecked());
+    gSettings->setCandidateWinHorizontal(ui->checkBoxCandidateWinHorizontal->isChecked());
 }
 
 void SettingsDialog::on_buttonBox_rejected()