+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * New feature Tab to switch between the matches in candidate window
 * While typing numbers in phonetic method, pressing : should make it : not ঃ
 * Old Reph feature in fixed typing method
+* User can select the orientation of candidate window in Settings
 
 ## 0.1.0
 Beta release